Use RS256 JWT encryption if rsa_private_key is present
Python Object is encrypted into byte-string using
RS256 algorithm and rsa_private_key from Variables.

Later this byte-string is converted to utf-8 and
set as payload field of json data that is POSTed.

Before POSTing json data is converted from utf-8
(hardcoded in Request package) to byte-string.

On the server side byte-string should be converted
to utf-8.

Payload field should be extracted and converted
from utf-8 to byte-string. This byte-string can be
used in jwt.decode with the public key.

The result of jwt.decode is byte-string. To display
it you should convert it to utf-8.