POST results when DagRun is "success"
Results field is added to the payload
only when the DagRun's state is "success".
For more details how the outputs are
retreived from CWLJobGatherer step, see
comments in post_status_info function.

Fix bug of posting payload if jwt is not
used (call decode in the correct place)

enter the commit message for your changes. Lines starting