Set correct default start_date for CWLDAG
Originally, every time the dag folder was loaded
start_date was set to utcnow. This value was also
used to set start_date for every task of the DAG.

When triggering DAG (creating DagRun for this DAG)
without specifying execution_date, the later one
will be set to utcnow by default.

Note, that DagRun will fail its verify_integrity test
if task.start_date > self.execution_date. In case we
use utcnow for both start_date and execution_date,
the order of setting this two variables matters.

If we use api_client set by default to
- airflow.api.client.local_client
each time we call "airflow trigger" we first load all
dags from dags folder (it means we set start_date
values to utcnow) and then we create DagRun with the
execution_date that has utcnow bigger then calculated one
for start_date. So we don't have any problems triggering
such dags.

BUT, if we want to use API or we set api_client to
- airflow.api.client.json_client
we will first calculate execution_date and then start_date.
Moreover, the timedelta between execution_date and start_date
will depend on how often scheduler parses dag folder.

So the best solution to make sure we can trigger CWLDAG
is to set start_date to days_ago(14). That means that if we
triggered our dag today and it was in a line for more then
14 days, it will have never been triggered.

To test triggering dags through API use trigger.py script.